Added setup.py Python script to transform the dbus-python project into a Python EGG.
This script transform dbus-python to be a regular Python EGG, which means that it can be integrated
as a dependency of any Python package, and be automatically downloaded by easy_install or other
Python deployment tools (like zc.buildout, etc.)

Currently, to build the egg, ones have to first "pre-configure" the dbus-python distribution by
running "./autogen.sh" or "./configure", which will generate some files needed by the compilation of
the package.
Then, you can use setup.py to build a binary egg, or a source egg (using bdist_egg or sdist).

Glib package will be compiled only if Glib development libraries are present on the system.